Move safeWaitForAsync() to the `testing/` folder
And also delete the karma-utils.spec.ts file, since we don't have unit
tests for the files in the `testing/` folder.
