+++ client/src/app/users/add-user.component.spec.ts
@@ -8,7 +8,7 @@ import { MatSnackBarModule } from '@angular/material/snack-bar';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { RouterTestingModule } from '@angular/router/testing';
 import { MockUserService } from 'src/testing/user.service.mock';
-import { safeWaitForAsync } from '../utils/karma-utils';
+import { safeWaitForAsync } from '../../testing/karma-utils';
 import { AddUserComponent } from './add-user.component';
 import { UserService } from './user.service';
